make dropdown factors hierachical
the dropdown factors now follow the following logic: the second dropdown selector shows only options available given the choices of the first one. the third only the choices possible the first and second one and so on

For example: you have two factors: country and blood type. Let's say country gets selected from the first selector and is set to US. If only americans with blood types 0, A and B answered the survey, the second dropdown menu will only give you those levels in case US is selected in the first dropdown menu. This avoids "empty factor combinations"